Feat(save_values): Save current values when saving the configuration
Now the values of the sliders, texts and checkboxes get stored when "save to file" is used, and get loaded when "load from file" is called.

Note: This functionality only works for the widgets in the "widget folder", namely: bool, double, int, uint and string. I do not know if there are any more, but in that case the value will not be saved and the plugin will work as before.